Remove external references
External references get flooded when images are present. An image with a link to its attachment
created not only a reference to the image, but its attachment page. While this behaviour could
have been fixed with some crafty regexing, the external URL feature was not as important as the
internal cross-reference one, so it has been removed.