Make pages more like wiki pages
Page authors and post dates are now hidden from the admin view. A special sidebar is now displayed for
pages, which is intended to hold a "Page Contents" widget which will eventually generate the page
contents with clickable links. A special "Page Root" template is available which does not display
any sidebar, intended for the "front page" of the pages section of the website, which is itself a
giant contents page.