Fix prefetch unread flag bug arising in some browsers
Browsers with link prefetching enabled use the <link rel="next"> element that
WordPress defines in the HTML <head> section to prefetch the next page. This
has the unintended side-effect of setting that page as read even when the user
has not viewed it.

This commit disables the display of adjacent post links in the HTML <head>.
An alternative approach could be to detect these prefetch requests, but
currently browsers have not standardised the header key being used.
Furthermore, another approach could be to detect when the user actually
looks at a page using JavaScript, but this is circumvented by blocking of
JavaScript.