Remove manual WP_Query SQL modifications for including coauthors in search results
The system copied from the CoAuthors Plus plugin modified the SQL query in WP_Query
directly. This is discouraged and unnecessary since WP_Query can include taxonomy
terms in searches directly. The down-side is that WP_Query cannot match against
authors OR taxonomy terms, only AND. That means that the default WP_Query object
parsed from a request for an author page must be modified to remove the author
clause and then have only the taxonomy term for the author added. This breaks
WordPress functionality elsewhere, such as the handling of 404 pages when there
are no author posts found, so the author clause has to be restored once the
query has been executed.