Fix bug with terms on login
When logging in, users of blogs would have their coauthor term added to the primary blog (usually site ID 1)
because of the hook 'check_coauthor_term_on_login' which creates the coauthor term if it doesn't exist.
This meant that users of secondary blogs who were not users of the primary blog would still be available
for selection as authors on the new post screen of the primary blog. They would not however save correctly
because the users were not members of that blog; they would be silently discard. This change fixes the
creation of the coauthor terms to only create terms on the blogs the user is a member of.

Existing sites running ALP will still have these terms, however. The rebuild coauthor tool will be updated
to delete terms of users who are not members of the blog.